+++ src/functions/index.ts
@@ -11,6 +11,16 @@ import {
   EmulatorStopForPubSub,
   EmulatorStopForSpanner,
 } from './emulator/index.js';
+import {
+  EventTopicBrokerCreateTopicForPubSub,
+  EventTopicBrokerCreateTriggerForCloudRun,
+  EventTopicBrokerDeleteTopicForPubSub,
+  EventTopicBrokerDeleteTriggerResourceForCloudRunInvokerRole,
+  EventTopicBrokerDeleteTriggerResourceForPubSubSubscription,
+  EventTopicBrokerDeleteTriggerResourceForServiceAccount,
+  EventTopicBrokerGetTopicIdForPubSub,
+  EventTopicBrokerPublishEventsForGoogle,
+} from './event-topic/index.js';
 import { GoogleAppCheckGenerateToken } from './google-app-check/index.js';
 import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage/index.js';
 import { GoogleFirestoreMergeRules } from './google-firestore/index.js';
@@ -43,6 +53,14 @@ export function registerFunctions(context: ModuleRegistrationContext) {
     EmulatorStopForIdentityPlatform,
     EmulatorStopForPubSub,
     EmulatorStopForSpanner,
+    EventTopicBrokerCreateTopicForPubSub,
+    EventTopicBrokerCreateTriggerForCloudRun,
+    EventTopicBrokerDeleteTopicForPubSub,
+    EventTopicBrokerDeleteTriggerResourceForCloudRunInvokerRole,
+    EventTopicBrokerDeleteTriggerResourceForPubSubSubscription,
+    EventTopicBrokerDeleteTriggerResourceForServiceAccount,
+    EventTopicBrokerGetTopicIdForPubSub,
+    EventTopicBrokerPublishEventsForGoogle,
     GoogleAppCheckGenerateToken,
     GoogleFirebaseStorageMergeRules,
     GoogleFirestoreMergeRules,