+++ src/services/index.ts
@@ -1,3 +1,4 @@
+export { CloudRunPubSubTriggerService } from './cloud-run-pubsub-trigger.js';
 export { CloudRunService } from './cloud-run.js';
 export * from './firebase-app.errors.js';
 export { FirebaseAppService, FirebaseAppType } from './firebase-app.js';